Fix lose ACL because master deletes tmp dir when start
Reviewers: #hbase, huzheng

Reviewed By: #hbase, huzheng

Maniphest Tasks: T14456

Differential Revision: https://phabricator.d.xiaomi.net/D163172

+++ hbase-server/src/main/java/org/apache/hadoop/hbase/master/MasterFileSystem.java
@@ -526,14 +526,18 @@ public class MasterFileSystem {
           HFileArchiver.archiveRegion(fs, this.rootdir, tabledir, regiondir);
         }
       }
-      if (!fs.delete(tmpdir, true)) {
-        throw new IOException("Unable to clean the temp directory: " + tmpdir);
+      if (!conf.getBoolean(HConstants.HDFS_ACL_ENABLE, false)) {
+        if (!fs.delete(tmpdir, true)) {
+          throw new IOException("Unable to clean the temp directory: " + tmpdir);
+        }
       }
     }
 
-    // Create the temp directory
-    if (!fs.mkdirs(tmpdir)) {
-      throw new IOException("HBase temp directory '" + tmpdir + "' creation failure.");
+    if (!fs.exists(tmpdir)) {
+      // Create the temp directory
+      if (!fs.mkdirs(tmpdir)) {
+        throw new IOException("HBase temp directory '" + tmpdir + "' creation failure.");
+      }
     }
   }
